update_0325
TODO Make Game 0325:

[Windows API를 활용한 2D 게임 제작]
기간: 26-03-16 ~ 26-04-13

[Todo List - Content]
-프로젝트 스크립트 중 1줄짜리 Get, Set 제외 모든 .h에 구현된 함수 .cpp로 옮기기
-삼각뿔 모양의 Mesh 생성
-배경 이동방식 설계 (청크 기반 생성 or 시차 스크롤링) (todo)
-Camera Component 2개 이상 배치 시, 투영계산에서 크래시 나는 현상 원인 알아보기
-GameObject 복사 생성 시, 콘텐츠 Script가 있다면, 멤버에 따라 복사생성자 구현이 필요한지 확인하기
-enemy 상태를 관리하는 클래스를 설계하여 mon1 Prefab 제작하기
-Mesh 모양에 맞춘 Collider 구현하기
-Player Jump 자연스럽게 구현하기

[Todo List - ImGui]
-Asset UI 구현
 -> GraphicShaderUI, MeshUI, PrefabUI, SoundUI 남음
-현재 만들어놓은 콘텐츠 스크립트, 멤버 UI 표시 기능 구현하기
-EScriptUI 파라미터 중, 값 변경을 시도하면, 원본 주소에 접근해 변경한 값을 세팅하는 기능 구현하기

[Finish Work]
-Player 행동에 따라 다른 Flipbook 재생 구현 (todo)
 -> 근접 공격 상태 클래스를 받아, Punch 기능과 Flipbook 재생 완료하기
 -> Punch Flipbook의 FPS를 높게 설정하고, Flipbook 재생이 끝나면 Idle로 돌아가게 설정, 추후 공격 중임을 나타내는 bool 변수와 Flipbook 타이밍 맞추거나, 다른 방식으로 공격 설계
-Player 이동에 가속도 추가 (todo)
-Player 이동과 추락상태 전환 자연스럽게 수정하기 (todo)
 -> 부모 클래스에 가속도를 받아 추락하는 ApplyGravity 함수를 만들고, 필요한 파생 클래스에서 호출
-Flipbook Index에 따라 Play FPS 매개변수 다르게 전달하게 변경하기

+++ DirectX11_Engine/GameClient/CFlipbookRender.cpp
@@ -194,7 +194,7 @@ void CFlipbookRender::SaveToLevelFile(FILE* _File)
 
 	fwrite(&m_CurFlipbook, sizeof(int), 1, _File);
 	fwrite(&m_CurSprite, sizeof(int), 1, _File);
-	fwrite(&m_FPS, sizeof(int), 1, _File);
+	fwrite(&m_FPS, sizeof(float), 1, _File);
 	fwrite(&m_RepeatCount, sizeof(int), 1, _File);
 }
 
@@ -213,6 +213,6 @@ void CFlipbookRender::LoadFromLevelFile(FILE* _File)
 
 	fread(&m_CurFlipbook, sizeof(int), 1, _File);
 	fread(&m_CurSprite, sizeof(int), 1, _File);
-	fread(&m_FPS, sizeof(int), 1, _File);
+	fread(&m_FPS, sizeof(float), 1, _File);
 	fread(&m_RepeatCount, sizeof(int), 1, _File);
 }
